+++ spec/controllers/tokens_controller_spec.rb
@@ -106,12 +106,13 @@ describe Doorkeeper::TokensController do
 
       it "should call :before_successful_authorization callback" do
         expect(Doorkeeper.configuration)
-          .to receive_message_chain(:before_successful_authorization, :call).with(instance_of(described_class))
+          .to receive_message_chain(:before_successful_authorization, :call).with(instance_of(described_class), nil)
       end
 
       it "should call :after_successful_authorization callback" do
         expect(Doorkeeper.configuration)
-          .to receive_message_chain(:after_successful_authorization, :call).with(instance_of(described_class))
+          .to receive_message_chain(:after_successful_authorization, :call)
+          .with(instance_of(described_class), instance_of(Doorkeeper::OAuth::Hooks::Context))
       end
     end
 
@@ -126,7 +127,7 @@ describe Doorkeeper::TokensController do
 
       it "should call :before_successful_authorization callback" do
         expect(Doorkeeper.configuration)
-          .to receive_message_chain(:before_successful_authorization, :call).with(instance_of(described_class))
+          .to receive_message_chain(:before_successful_authorization, :call).with(instance_of(described_class), nil)
       end
 
       it "should not call :after_successful_authorization callback" do