+++ spec/dummy/app/helpers/application_helper.rb
@@ -2,6 +2,6 @@
 
 module ApplicationHelper
   def current_user
-    @current_user ||= User.find_by_id(session[:user_id])
+    @current_user ||= User.find_by(id: session[:user_id])
   end
 end